Implement responsive layout for recently added items
Moves browse item out to its own file
