+++ assets/css/app.css
@@ -39,6 +39,14 @@
   visibility: visible;
 }
 
+.recent-container {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  grid-template-rows: 1fr 0px 0px;
+  @apply gap-x-2;
+
+}
+
 /*
  * Make LiveView wrapper divs transparent for layout.
  * This makes it possible to use LiveViews as flex children for example.
@@ -62,6 +70,11 @@
 
 @utility btn-secondary {
   @apply cursor-pointer uppercase border bg-sage text-black;
+}
+
+.btn-arrow {
+  clip-path: polygon(0% 0%, 100% 50%, 0% 100%);
+  @apply bg-dark-blue h-14 cursor-pointer border;
 }
 
 @utility active {
